[NO CODE MODIFICATIONS] split up Rust codegen in submodules
