+++ crates/re_types/source_hash.txt
@@ -1,4 +1,4 @@
 # This is a sha256 hash for all direct and indirect dependencies of this crate's build script.
 # It can be safely removed at anytime to force the build script to run again.
 # Check out build.rs to see how it's computed.
-d52b15386c820be3e4b12f920c9d589b7b4298f113c7ec5614f6e71e3b6235d1
+dd0a9d78997443b26b2b3ed294c09aee58b285b55c42edaad8479b587df0b394
