+++ crates/re_types_builder/src/codegen/rust/mod.rs
@@ -1,0 +1,9 @@
+//! Implements the Rust codegen pass.
+
+mod api;
+mod arrow;
+mod deserializer;
+mod serializer;
+mod util;
+
+pub use self::api::RustCodeGenerator;
